Move duplicated paths
