+++ app/grandchallenge/pages/urls.py
@@ -2,7 +2,6 @@ from django.urls import path
 
 from grandchallenge.pages.views import (
     ChallengeHome,
-    FaviconView,
     PageCreate,
     PageDelete,
     PageDetail,
@@ -15,32 +14,6 @@ app_name = "pages"
 urlpatterns = [
     path("pages/", PageList.as_view(), name="list"),
     path("pages/create/", PageCreate.as_view(), name="create"),
-    # Favicons
-    path(
-        "favicon.ico/",
-        FaviconView.as_view(rel="shortcut icon"),
-        name="favicon",
-    ),
-    path(
-        "apple-touch-icon.png/",
-        FaviconView.as_view(rel="apple-touch-icon"),
-        name="apple-touch-icon",
-    ),
-    path(
-        "apple-touch-icon-precomposed.png/",
-        FaviconView.as_view(rel="apple-touch-icon-precomposed"),
-        name="apple-touch-icon-precomposed",
-    ),
-    path(
-        "apple-touch-icon-<int:size>x<int>.png/",
-        FaviconView.as_view(rel="apple-touch-icon"),
-        name="apple-touch-icon-sized",
-    ),
-    path(
-        "apple-touch-icon-<int:size>x<int>-precomposed.png/",
-        FaviconView.as_view(rel="apple-touch-icon-precomposed"),
-        name="apple-touch-icon-precomposed-sized",
-    ),
     path("", ChallengeHome.as_view(), name="home"),
     path("<slug:page_title>/", PageDetail.as_view(), name="detail"),
     path("<slug:page_title>/update/", PageUpdate.as_view(), name="update"),
