+++ app/grandchallenge/pages/views.py
@@ -5,10 +5,8 @@ from django.views.generic import (
     DeleteView,
     DetailView,
     ListView,
-    RedirectView,
     UpdateView,
 )
-from favicon.models import Favicon
 
 from grandchallenge.core.permissions.mixins import (
     UserAuthAndTestMixin,
@@ -115,30 +113,3 @@ class PageDelete(
     def delete(self, request, *args, **kwargs):
         messages.success(self.request, self.success_message)
         return super().delete(request, *args, **kwargs)
-
-
-class FaviconView(RedirectView):
-    """
-    Some browsers do not follow the favicon links in base.html, so do this
-    explicitly here.
-    """
-
-    permanent = False
-    rel = "shortcut icon"
-
-    def get_redirect_url(self, *args, **kwargs):
-        fav = Favicon.objects.filter(isFavicon=True).first()
-
-        if not fav:
-            return None
-
-        if self.rel == "shortcut icon":
-            size = 32
-        else:
-            # This is the largest icon from
-            # https://github.com/audreyr/favicon-cheat-sheet
-            size = kwargs.get("size", 180)
-
-        default_fav = fav.get_favicon(size=size, rel=self.rel)
-
-        return default_fav.faviconImage.url
